Remove queriers that are no longer needed.
